Arm the noise sources independently from arm_rst
The noise sources were armed from the same signal used to arm the master
timebase (namely arm_rst).  Once armed, the noise sources resync on the
sync pulse from the master timebase's sync generator.  This creates a race
condition.  If the not-yet-synced sync generator outputs a non-syncronized sync
pulse after arming, but before receving an external sync pulse, the noise generator will reseed asynchrously with the external sync signal.

Using the arm_rst signal to reseed the noise generators also means that the
noise generators cannot be reseeded arbtrarily after the initial
synchronization without resetting the master counter which can cause mayehm
downstream.

To solve these problems, a previously unused bit (bit 15) of the feng_ctrl
register is now used solely to rearm the noise generators.

In order to ensure that multiple ROACHes can have their noise sources rearmed
by the same sync pulse, it is necessary for the control software to know when
the sync pulses will be occurring.  The 200e6/2^27 Hz frequency of the internal
sync_gen signal is not commensurate with 1 PPS so it is hard for external
software to know when the next sync_gen pulse is going to occur.  To avoid
that, we create an internal 1 PPS signal that is synchronized to the external
sync signal at the same time the sync_gen signal is synchronized with the
external sync signal.  This internal 1 PPS signal will have the same
relationship to mcnt on all synchronized ROACHes yet still allow the control
software to know (to the precision of NTP) when the noise generators will be
synchronized so that it can arm all of them between 1 PPS pulses.